requirements.txt
asgiref==3.5.2
astroid==2.11.7
atomicwrites==1.4.0
attrs==21.2.0
autopep8==1.7.0
certifi==2021.10.8
charset-normalizer==2.0.10
colorama==0.4.5
dill==0.3.5.1
distlib==0.3.4
Django==4.1.2
django-crispy-forms==1.14.0
django-environ==0.9.0
django-login-required-middleware==0.9.0
django-widget-tweaks==1.4.12
filelock==3.7.0
geojson==2.5.0
idna==3.3
iniconfig==1.1.1
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
packaging==21.0
platformdirs==2.5.2
pluggy==1.0.0
py==1.10.0
pycodestyle==2.9.1
pylint==2.14.5
pyowm==3.3.0
pyparsing==3.0.3
PySocks==1.7.1
pytest==6.2.5
requests==2.27.1
six==1.16.0
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.3
tqdm==4.62.3
typing_extensions==4.2.0
tzdata==2022.5
urllib3==1.26.8
virtualenv==20.14.1
wrapt==1.14.1
